requirements.txt
pandas>=2.0.0
numpy>=1.24.0
spacy>=3.7.2
vaderSentiment>=3.3.2
gensim>=4.3.2
streamlit>=1.28.0
google-api-python-client>=2.108.0
python-dotenv>=1.0.0
